+++ pointfreeco.playground/contents.xcplayground
@@ -7,7 +7,5 @@
         <page name='Signup Email'/>
         <page name='Team Invites'/>
         <page name='Account'/>
-        <page name='Email Tester'/>
-        <page name='Style Inliner Experiments'/>
     </pages>
 </playground>
